// src/bsdcpolls-frontend/src/app/generated/api.ts
export enum SurveyAnswerType {
    MultipleChoice = "MultipleChoice",
    ShortText = "ShortText",
    LongText = "LongText",
    DocumentUpload = "DocumentUpload",
}

export interface SurveyChoiceOption {
    uid: string;
    text: string;

    [key: string]: any;
}

export interface SurveyConditionalBranch {
    parentChoiceUid: string;
    questions: SurveyQuestionNode[];

    [key: string]: any;
}

export interface SurveyQuestionNode {
    uid: string;
    text: string;
    answerType: SurveyAnswerType;
    isRequired: boolean;
    choices?: SurveyChoiceOption[] | undefined;
    branches?: SurveyConditionalBranch[] | undefined;

    [key: string]: any;
}

export interface SurveyQuestionTreeDocument {
    questions: SurveyQuestionNode[];

    [key: string]: any;
}